move modules into subfolder
